git-svn-id: svn://svn.code.sf.net/p/win32-framework/code@2743 40094c20-7226-0410-9e7a-ca24b3fb783e

+++ samples/Explorer/src/MyTreeView.cpp
@@ -32,7 +32,7 @@ int CALLBACK CMyTreeView::CompareFunction(LPARAM param1, LPARAM param2, LPARAM)
     if(FAILED(result))
         return 0;
 
-    return (short)SCODE_CODE(GetScode(result));
+    return SCODE_CODE(GetScode(result));
 }
 
 // Identifies the tree view item for the point, and calls DoItemMenu.
